requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic-settings==2.6.1

# Database (SQLite for MVP)
sqlalchemy==2.0.36
alembic==1.14.0

# Agent Framework
langgraph==0.2.45
langchain-core==0.3.7
langchain-openai==0.2.8
langchain-google-genai==2.0.0

# Gmail
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.154.0

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
jinja2==3.1.4
